actualizacion 0823

+++ ft_intlen.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_intlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jperez-r <joserabasot@gmail.com>           +#+  +:+       +#+        */
+/*   By: jperez-r <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/18 22:01:53 by jperez-r          #+#    #+#             */
-/*   Updated: 2019/12/19 04:34:53 by jperez-r         ###   ########.fr       */
+/*   Updated: 2020/12/10 23:09:20 by jperez-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_intlen(int n)
+size_t	ft_intlen(long int n)
 {
 	size_t	len;
